requirements.txt
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
pandas
pillow
scikit_image
scikit_video
scipy
soundfile
torch
torchaudio
torchvision
tqdm
tslearn
omegaconf
pytorch-fid
einops
matplotlib
